fix: make bootstrap work

+++ bootstrap.sh
@@ -6,12 +6,21 @@ git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
 
 # install nvim
 curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
-rm -rf ~/nvim-linux64
+if [[ -f ~/nvim-linux64 ]]; then
+  echo "Removing previous installation..."
+  rm -rf ~/nvim-linux64
+fi
 tar -C ~ -xzf nvim-linux64.tar.gz
 
 read -p "Do you want to proceed adding nvim to your bashrc file? (Y/n): " answer
 if [[ $(echo "$answer" | tr '[:upper:]' '[:lower:]') == 'y' ]]; then
-  echo "export PATH=$PATH:~/nvim-linux64/bin" >> ~/.bashrc
+  if grep -q "nvim-preblock" ~/.bashrc;
+  then
+    echo "nvim path and aliases already configured!"
+    exit 0
+  fi
+  echo "# nvim-preblock" >> ~/.bashrc
+  echo "export PATH=\$PATH:~/nvim-linux64/bin" >> ~/.bashrc
   echo "alias vim=nvim" >> ~/.bashrc
   echo "alias vi=nvim" >> ~/.bashrc
   source ~/.bashrc
